% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placelabel.R
\name{plot.placelabel}
\alias{plot.placelabel}
\title{Plot place label}
\usage{
\method{plot}{placelabel}(x, labels = NULL, adj = NULL, pos = NULL,
  offset = 0.5, vfont = NULL, cex = 1, col = NULL, font = NULL, ...,
  add = FALSE)
}
\arguments{
\item{x}{placelabel}

\item{labels}{a character vector or \link{expression} specifying
    the \emph{text} to be written.  An attempt is made to coerce other
    language objects (names and calls) to expressions, and vectors and
    other classed objects to character vectors by \code{\link{as.character}}.
    If \code{labels} is longer than \code{x} and
    \code{y}, the coordinates are recycled to the length of \code{labels}.}

\item{adj}{one or two values in \eqn{[0, 1]} which specify the x
    (and optionally y) adjustment of the labels.  On most devices values
    outside that interval will also work.}

\item{pos}{a position specifier for the text.  If specified this
    overrides any \code{adj} value given.  Values of \code{1},
    \code{2}, \code{3} and \code{4}, respectively indicate
    positions below, to the left of, above and to the right of
    the specified coordinates.}

\item{offset}{when \code{pos} is specified, this value gives the
    offset of the label from the specified coordinate in fractions
    of a character width.}

\item{vfont}{\code{NULL} for the current font family, or a character
    vector of length 2 for Hershey vector fonts.  The first element of
    the vector selects a typeface and the second element selects a
    style.  Ignored if \code{labels} is an expression.}

\item{cex}{numeric \bold{c}haracter \bold{ex}pansion factor; multiplied
    by \code{\link{par}("cex")} yields the final character size.
    \code{NULL} and \code{NA} are equivalent to \code{1.0}.}

\item{col}{the color and (if \code{vfont = NULL}) font to be
    used, possibly vectors.  These default to the values of the global
    \link{graphical parameters} in \code{\link{par}()}.}

\item{font}{the color and (if \code{vfont = NULL}) font to be
    used, possibly vectors.  These default to the values of the global
    \link{graphical parameters} in \code{\link{par}()}.}

\item{...}{arguments passed to `graphics::plot` (ignored if `add = TRUE`)}

\item{add}{create a new plot if `FALSE` (the default), or add to an existing plot}
}
\value{
nothing
}
\description{
Plot a `placelabel` either by creating a blank plot and plotting only labels (the default),
or add to an existing plot with `add = TRUE`.
}
\details{
This plot method is a thin wrapper around `graphics::text` and includes all of those arguments.
The `...` argument is passed to `graphics::plot` to control the set up of the plot
when `add = FALSE`.
}
\examples{
library(sf)
north_carolina <- read_sf(system.file("shape/nc.shp", package="sf"))
nc_labs <- placelabel(north_carolina)
plot(nc_labs)
nc_labs <- placelabel(north_carolina, text = "NAME")
plot(nc_labs, cex = 0.5)
## leverage sf aesthetics without destroying plot layout
plot(north_carolina[1], add = TRUE)
plot(nc_labs, cex = 0.5, add = TRUE)
plot(st_geometry(north_carolina))
plot(nc_labs, "BIR74", add = TRUE, cex = 0.5)
}
