% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placelabel.R
\name{placelabel}
\alias{placelabel}
\alias{placelabel}
\alias{placelabel.default}
\alias{placelabel}
\alias{placelabel.sf}
\title{Place labels}
\usage{
placelabel(x, ...)

\method{placelabel}{default}(x, ...)

\method{placelabel}{sf}(x, ...)
}
\arguments{
\item{x}{a spatial object, sf or anything convertible with `sf::st_as_sf` supported}

\item{...}{arguments passed to `graphics::plot`,}
}
\description{
A place label is a simple data structure that exists to plot
spatial data as labels.
}
\examples{
library(sf)
north_carolina <- read_sf(system.file("shape/nc.shp", package="sf"))
nc_labs <- placelabel(north_carolina)
plot(st_geometry(north_carolina))
plot(nc_labs, add = TRUE, cex = 0.8)
}
